// ServerlessFramework/src/entities/enums/product.ts
export enum ProductNameEnum {
  UnlimitedPlans = 'Unlimited plans',
  AirTime = 'Airtime',
  SimValidity = 'Sim validity',
  DoorDelivery = 'Door Delivery',
}

export enum ProductSlugEnum {
  UnlimitedPlans = 'unlimitedplans',
  AirTime = 'airtime',
  SimValidity = 'simvalidity',
  DoorDelivery = 'doordelivery',
}